debug server

+++ codes/server/server/dataConnect/ClientThread.rs
@@ -10,7 +10,6 @@ use std::fs::File;
 use super::super::database::Query::Query;
 use super::super::database::Query::FileItem;
 use super::super::database::Query::RequestItem;
-use super::FileTransporter;
 
 /*在crate root 中声明 "extern crate chrono;"
 cargo.toml中增加：
@@ -42,7 +41,7 @@ impl ClientThread{
         }
     }
 
-    pub fn run(&mut self){
+    pub fn run(mut self){
         let mut status:bool = false;
         println!("start!");
         //这两行java代码未对应实现
@@ -76,14 +75,14 @@ impl ClientThread{
         }
     }
 
-    pub fn recv_required_fragment(&self)->bool{
+    pub fn recv_required_fragment(mut self)->bool{
         let mut status:bool = true;
         let command:Vec<&str> = self.sentence[..].split(' ').collect();
         let id:i32 = command[1].parse().unwrap();
         let fid:i32 = command[2].parse().unwrap();
 
         let query = Query::new();
-        let request = query.queryRequest_Byid(id);
+        let mut request = query.queryRequest_Byid(id);
 
         if request.get_fragment_id() != fid || request.get_type() != 1{
             self.client_socket.write(b"ERROR!\n");
@@ -91,11 +90,13 @@ impl ClientThread{
             status = false;
         }
         else{
-            let s = self.download_folder_path.into_os_string().into_string().unwrap() + &fid.to_string();
+            let mut s: String = self.download_folder_path.into_os_string().into_string().unwrap();
+            s.push_str(&fid.to_string());
+                //+ &fid.to_string();
             let recv_file = File::create(s).unwrap();
             self.client_socket.write(b"received!\n");
             self.client_socket.flush();
-            status = Filetransporter::recv_file(recv_file, &self.client_socket);
+            status = super::FileTransporter::recv_file(recv_file, &self.client_socket);
             if status {
                 self.client_socket.write(b"received!\n");
                 self.client_socket.flush();
@@ -106,20 +107,21 @@ impl ClientThread{
         status
     }
 
-    pub fn send_fragment(&self)->bool{
+    pub fn send_fragment(mut self)->bool{
         let mut status:bool = true;
         let command:Vec<&str> = self.sentence[..].split(' ').collect();
         let id:i32 = command[1].parse().unwrap();
         let fid:i32 = command[2].parse().unwrap();
         
         let query = Query::new();
-        let request = query.queryRequest_Byid(id);
+        let mut request = query.queryRequest_Byid(id);
 
         if request.get_fragment_id() != fid || request.get_type() != 2 {
             status = false;
         }
         else{
-            let s = self.upload_folder_path.into_os_string().into_string().unwrap() + &fid.to_string();//upload_folder_path的形式还未确定
+            let mut s: String = self.download_folder_path.into_os_string().into_string().unwrap();
+            s.push_str(&fid.to_string());
             let send_file = File::open(s);
             match send_file{
                 Err(e) => {
@@ -130,15 +132,19 @@ impl ClientThread{
                     status = super::FileTransporter::send_file(file, &self.client_socket);
                     if status{
                         let mut in_from_cilent = BufReader::new(self.client_socket);
-                        let sentence = String::new();
+                        let mut sentence = String::new();
                         in_from_cilent.read_line(&mut sentence).unwrap();
-                        match sentence{
-                            "received!".to_string() => {
-                                //sendFile.delete();
-                                query.deleteRequest(request.get_id());
-                                //query.alterFragment(fid, Integer.toString(request.getDeviceId()));
-                                query.alterFragment(fid, request.get_device_id().to_string());
-                            }
+                        let re = vec!['r', 'e', 'c', 'e', 'i', 'v', 'e', 'd', '!'];
+                        let mut n: usize = 0;
+                        for sen in sentence.chars() {
+                            if sen != re[n] {break;}
+                            else {n = n + 1;}
+                        }
+                        if n == re.len() - 1 {
+                            //sendFile.delete();
+                            query.deleteRequest(request.get_id());
+                            //query.alterFragment(fid, Integer.toString(request.getDeviceId()));
+                            query.alterFragment(fid, request.get_device_id().to_string());
                         }
                     }
                 }
@@ -149,14 +155,14 @@ impl ClientThread{
         status
     }
 
-    pub fn delete_fragment(&self)->bool{
+    pub fn delete_fragment(&mut self)->bool{
         let mut status:bool = true;
         let command:Vec<&str> = self.sentence[..].split(' ').collect();
         let id:i32 = command[1].parse().unwrap();
         let fid:i32 = command[2].parse().unwrap();
 
         let query = Query::new();
-        let request = query.queryRequest_Byid(id);
+        let mut request = query.queryRequest_Byid(id);
 
         if request.get_fragment_id() != fid || request.get_type() != 3 {
             self.client_socket.write(b"ERROR!\n");
@@ -173,7 +179,7 @@ impl ClientThread{
         status
     }
 
-    pub fn register_file(&self)->bool{
+    pub fn register_file(&mut self)->bool{
         let command:Vec<&str> = self.sentence[..].split(' ').collect();
         let noa:i32 = command[5].parse().unwrap();
         let isf:bool = command[6].parse().unwrap();
@@ -196,7 +202,7 @@ impl ClientThread{
         true
     }
 
-    pub fn recv_file_fragment(&self)->bool{
+    pub fn recv_file_fragment(mut self)->bool{
         let mut status: bool = true;
         let command:Vec<&str> = self.sentence[..].split(' ').collect();
         let file_id:i32 = command[1].parse().unwrap();
@@ -204,7 +210,7 @@ impl ClientThread{
         let fragment_count:i32 = command[3].parse().unwrap();
 
         let query = Query::new();
-        let file = query.queryFile_Byid(file_id);
+        let mut file = query.queryFile_Byid(file_id);
 
         if file.get_noa() != -1 * fragment_count || fragment_num >= fragment_count || fragment_num < 0 {
             self.client_socket.write(b"ERROR!\n");
@@ -213,12 +219,14 @@ impl ClientThread{
         }
         else{
             let temp = file_id * 100 + fragment_num;
-            let s:String = self.upload_folder_path.into_os_string().into_string().unwrap() + &temp.to_string();
+            let mut s: String = self.upload_folder_path.into_os_string().into_string().unwrap();
+            let mut s1: String = s.clone();
+            s.push_str(&temp.to_string());
             let recv_file = File::create(s).unwrap();
             self.client_socket.write(b"received!\n");
             self.client_socket.flush();
 
-            status = FileTransporter::recv_file(recv_file, &self.client_socket);
+            status = super::FileTransporter::recv_file(recv_file, &self.client_socket);
             if status{
                 query.addFragment(temp, "-1".to_string());
                 if fragment_num == fragment_count - 1 {
@@ -236,9 +244,8 @@ impl ClientThread{
                         for i in 0..fragment_count{
                             if query.deleteFragment(file_id * 100 + i) == 1 {
                                 let temp_2:i32 = file_id * 100 + i;
-                                let s:String = self.upload_folder_path.into_os_string().into_string().unwrap() 
-                                    + &temp_2.to_string();
-                                let f = File::create(s).unwrap();
+                                s1.push_str(&temp_2.to_string());
+                                let f = File::create(s1).unwrap();
                             }
                         }
                     }
@@ -253,7 +260,7 @@ impl ClientThread{
         status
     }
 
-    pub fn check_folder(&self)->bool{
+    pub fn check_folder(mut self)->bool{
         let command:Vec<&str> = self.sentence[..].split(' ').collect();
         let num:i32 = command[2].parse().unwrap();
 
@@ -268,7 +275,7 @@ impl ClientThread{
             in_from_client.read_line(&mut input).unwrap();
             let ipt = input.as_mut_vec();
             let input_vec:Vec<&str> = input[..].split(' ').collect();
-            let file = query.queryFile_Bypathname(Some(ipt[0].to_string()), Some(ipt[1].to_string()));
+            let mut file = query.queryFile_Bypathname(Some(ipt[0].to_string()), Some(ipt[1].to_string()));
             if  -1 == file.get_id() {
                 let dt = Local::today();
                 let mut date:String = dt.to_string();
@@ -303,43 +310,44 @@ impl ClientThread{
         true
     }
 
-    pub fn confirm(&self, id:&i32, num:&i32)->i32{
+    pub fn confirm(&mut self, id:&i32, num:&i32)->i32{
         let query = Query::new();
         let mut return_val:i32 = 0;
 
-        let di = query.queryOnlineDevice();
+        let mut di = query.queryOnlineDevice();
         //假定di类型为Vec<DeviceItem>
         if di.is_empty() {
             return -1;
         }
 
         let s = di.len();
+        let size: i32 = s as i32;
         if num <= &size {
-            let t: i32 = rand::thread_rng().gen_range(0, size).try_into().unwrap();
+            let t: i32 = rand::thread_rng().gen_range(0, size);
             for i in 0..*num{
-                let n: i32 = i.try_into().unwrap();
-                let temp = RequestItem::init_2(2, id * 100 + n, di[(n + t) % size].get_id());
+                let n: i32 = i as i32;
+                let temp = RequestItem::init_2(2, id * 100 + n, di[((n + t) % size) as usize].get_id());
                 query.addRequest(temp);
             }
         }
         else{
             let mut n:Vec<i32> = Vec::new();
-            let temp = num / (size as i32);
+            let temp = num / size;
             for i in 0..size {
                 n.push(temp);
             }
-            let m = num % (size as i32);
+            let m = num % size;
 
             let mut t = rand::thread_rng().gen_range(0, size);
             for i in 0..m {
-                n[t % size] = n[t % size] + 1;
+                n[(t % size) as usize] = n[(t % size) as usize] + 1;
                 t = t + 1;
             }
 
             let mut k:i32 = 0;
             for i in 0..size {
-                for j in 0..n[i] as usize{
-                    let temp = RequestItem::init_2(2, id * 100 + (i as i32), di[i].get_id());
+                for j in 0..n[i as usize] as usize{
+                    let temp = RequestItem::init_2(2, id * 100 + (i as i32), di[i as usize].get_id());
                     query.addRequest(temp);
                     k = k + 1;
                 }
